refactored and working on docker
