+++ .gitignore
@@ -1,3 +1,8 @@
-auth.py
 __pycache__
+
+#auth info
+auth.py
 leetcode_cookies.csv
+
+#debugging
+screenshots/
